+++ usecase/src/main/kotlin/dev/agner/portfolio/usecase/bond/position/model/BondPosition.kt
@@ -2,12 +2,13 @@ package dev.agner.portfolio.usecase.bond.position.model
 
 import dev.agner.portfolio.usecase.bond.model.Bond
 import kotlinx.datetime.LocalDate
+import java.math.BigDecimal
 
 data class BondPosition(
     val bond: Bond,
     val date: LocalDate,
-    val principal: Double,
-    val yield: Double,
-    val taxes: Double,
-    val result: Double,
+    val principal: BigDecimal,
+    val yield: BigDecimal,
+    val taxes: BigDecimal,
+    val result: BigDecimal,
 )
